configure: with bundled candl, add piplib to include path
Clay transitively includes piplibMP.h, but when candl is bundled, the
autotools build system cannot find it.  This is due to it being located
under ./candl/piplib/include that is present in the include path.  Fix
that.

The problem did not appear with "system" builds, most often used,
because all tools usually get installed into the same location, e.g.,
/usr/local/include.